+++ packages/cloud-cognitive/src/components/APIKeyModal/APIKeyModal.js
@@ -14,6 +14,7 @@ import {
   ModalFooter,
   ModalBody,
   TextInput,
+  PasswordInput,
   InlineLoading,
   Form,
   Button,
@@ -201,7 +202,7 @@ export let APIKeyModal = forwardRef(
             <>
               {body && <p className={`${blockClass}__body`}>{body}</p>}
               {!editing && apiKey && hasAPIKeyVisibilityToggle && (
-                <TextInput.PasswordInput
+                <PasswordInput
                   value={apiKey}
                   labelText={apiKeyLabel}
                   id={apiKeyInputId.current}