+++ packages/security/src/components/Header/HeaderListItem/__tests__/HeaderListItem-test.js
@@ -31,17 +31,5 @@ describe('HeaderListItem', () => {
     it('renders extra classes', () => {
       expect(headerListItem.hasClass(className)).toEqual(true);
     });
-
-    it('renders the `aria-expanded` attribute', () => {
-      expect(headerListItem.prop('aria-expanded')).toEqual(isExpanded);
-      headerListItem.setProps({ isExpanded: !isExpanded });
-      expect(headerListItem.prop('aria-expanded')).toEqual(!isExpanded);
-    });
-
-    it('renders the `aria-haspopup` attribute', () => {
-      expect(headerListItem.prop('aria-haspopup')).toEqual(hasPopup);
-      headerListItem.setProps({ hasPopup: !hasPopup });
-      expect(headerListItem.prop('aria-haspopup')).toEqual(!hasPopup);
-    });
   });
 });