+++ packages/security/src/components/Header/HeaderListItem/constants.js
@@ -10,24 +10,16 @@ import { namespace as headerNamespace } from '../constants';
 
 const defaultProps = {
   className: '',
-  hasPopup: false,
-  isExpanded: false,
 };
 
 const namespace = appendComponentNamespace(headerNamespace, 'list__item');
 
 const propTypes = {
-  /** @type {boolean|HTMLElement} List item children. */
-  children: PropTypes.oneOfType([PropTypes.bool, PropTypes.element]).isRequired,
+  /** @type {node} List item children. */
+  children: PropTypes.node.isRequired,
 
   /** @type {string} Extra classes. */
   className: PropTypes.string,
-
-  /** @type {boolean} Whether it has a popover or not. */
-  hasPopup: PropTypes.bool,
-
-  /** @type {boolean} Whether it is expanded or not. */
-  isExpanded: PropTypes.bool,
 };
 
 export { defaultProps, namespace, propTypes };