+++ packages/security/src/components/LayoutModules/DescriptionListModule/DescriptionListModule-story.js
@@ -28,6 +28,12 @@ export default {
     ...getDocsParameters(),
   },
 };
+// Overrides carbon's structuredListCell padding-left
+// See: https://www.carbondesignsystem.com/components/structured-list/usage/
+
+const style = {
+  paddingLeft: '0',
+};
 
 export const Default = () => (
   <DescriptionListModule>
@@ -73,7 +79,7 @@ export const Variant = () => (
     <TypeLayout>
       <TypeLayoutBody>
         <TypeLayoutRow>
-          <TypeLayoutCell>Term 1</TypeLayoutCell>
+          <TypeLayoutCell style={style}>Term 1</TypeLayoutCell>
 
           <TypeLayoutCell>
             Lorem ipsum dolor sit amet, consectetur adipiscing elit.
@@ -81,7 +87,7 @@ export const Variant = () => (
         </TypeLayoutRow>
 
         <TypeLayoutRow>
-          <TypeLayoutCell>Term 2</TypeLayoutCell>
+          <TypeLayoutCell style={style}>Term 2</TypeLayoutCell>
 
           <TypeLayoutCell>
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam
@@ -92,7 +98,7 @@ export const Variant = () => (
         </TypeLayoutRow>
 
         <TypeLayoutRow>
-          <TypeLayoutCell>Term 3</TypeLayoutCell>
+          <TypeLayoutCell style={style}>Term 3</TypeLayoutCell>
 
           <TypeLayoutCell>
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam