+++ packages/security/src/components/LayoutModules/TitleBarModule/index.js
@@ -4,7 +4,7 @@
  */
 
 import classnames from 'classnames';
-import { bool, elementType, node } from 'prop-types';
+import { bool, elementType, node, oneOf, string } from 'prop-types';
 import React, { createElement } from 'react';
 
 import LayoutModule, {
@@ -17,22 +17,39 @@ const namespace = 'title-bar';
 /**
  * Title bar modules provide interchangeable and reliable headings for establishing consistent hierarchies.
  */
-const TitleBarModule = ({ children, element, subsection, title, ...other }) => (
+const TitleBarModule = ({
+  children,
+  element,
+  subsection,
+  title,
+  label,
+  labelDirection,
+  ...other
+}) => (
   <LayoutModule
     className={classnames({
       [`${layoutModuleNamespace}--${namespace}--subsection`]: subsection,
+      [`${layoutModuleNamespace}--${namespace}--label`]: label,
     })}
     namespace={namespace}
     {...other}>
-    {createElement(
-      element,
-      {
-        ...getLayoutModuleProps({
-          namespace: `${namespace}__title`,
-        }),
-      },
-      title
-    )}
+    <div className={`${layoutModuleNamespace}--${namespace}__label-wrapper`}>
+      {label && (
+        <LayoutModule
+          namespace={`${namespace}__label order--${labelDirection}`}>
+          {label}
+        </LayoutModule>
+      )}
+      {createElement(
+        element,
+        {
+          ...getLayoutModuleProps({
+            namespace: `${namespace}__title`,
+          }),
+        },
+        title
+      )}
+    </div>
 
     {children && (
       <LayoutModule namespace={`${namespace}__items`}>{children}</LayoutModule>
@@ -44,8 +61,11 @@ TitleBarModule.propTypes = {
   /** Provide the items for the `TitleBarModule` */
   children: node,
 
-  /** Specify the base element to use to build the title */
-  element: elementType,
+  /** Provide the label for the `TitleBarModule` */
+  label: string,
+
+  /** Provide the labels direction for the `TitleBarModule` */
+  labelDirection: oneOf(['top', 'bottom']),
 
   /** Specify whether a subsection should be used */
   subsection: bool,
@@ -58,6 +78,8 @@ TitleBarModule.defaultProps = {
   children: null,
   subsection: false,
   element: 'h2',
+  label: null,
+  labelDirection: 'top',
 };
 
 export default TitleBarModule;