+++ packages/security/src/components/PageLayouts/stories/index.js
@@ -3,8 +3,9 @@
  * @copyright IBM Security 2020 - 2021
  */
 
-import { Grid } from 'carbon-components-react';
 import { layout04 } from '@carbon/layout';
+
+import { Grid } from 'carbon-components-react';
 import classnames from 'classnames';
 import React from 'react';
 
@@ -21,6 +22,7 @@ export default {
   parameters: {
     ...disableCentered(),
   },
+
   decorators: [
     (story) => (
       <>
@@ -53,12 +55,13 @@ export default {
 };
 
 // TODO: Remove workaround for https://github.ibm.com/security/design-core-experience/issues/241
-export const withContainer =
-  (WrappedComponent) =>
-  ({ className, ...other }) =>
-    (
+export const withContainer = function WithContainer(WrappedComponent) {
+  return function wrappedComponent({ className, ...other }) {
+    return (
       <WrappedComponent
         className={classnames('container--narrow', className)}
         {...other}
       />
     );
+  };
+};